move cross chain call parsing to helper
